requirements.txt
argparse
paho-mqtt